+++ src/main/java/com/scubakay/zombiescantgather/command/ZombiesCantGatherCommand.java
@@ -15,7 +15,7 @@ import net.minecraft.server.command.ServerCommandSource;
 
 import static com.scubakay.zombiescantgather.ZombiesCantGather.MOD_CONFIG;
 
-public class ZombiesCantGatherCommand extends RootCommand {
+public class ZombiesCantGatherCommand {
     private static final SuggestionProvider<ServerCommandSource> LISTED_ITEM_SUGGESTIONS = (context, builder) -> {
         for (String s : MOD_CONFIG.zombiesCantGather.get()) builder.suggest(s);
         return builder.buildFuture();
@@ -25,7 +25,7 @@ public class ZombiesCantGatherCommand extends RootCommand {
         LiteralCommandNode<ServerCommandSource> zombie = CommandManager
                 .literal("zombie")
                 .build();
-        getRoot().addChild(zombie);
+        RootCommand.getRoot().addChild(zombie);
 
         zombie.addChild(CommandUtil.getAddNode(
                 registryAccess,